refactor #260518: 招待コード生成コンポーネント分割
